requirements.txt
pydantic
langchain
langchain-openai
langchain-groq
langchain-anthropic
langchain-core
langchain-community
anthropic
defusedxml
pypdf
pymupdf
bs4
arxiv
wikipedia
python-dotenv
langchain_google_genai
pandas
langchain-pinecone
youtube_search
duckduckgo-search
yfinance
langgraph
fastapi
uvicorn
jinja2 
python-multipart
